+++ Events/DockedEvent.cs
@@ -23,6 +23,7 @@ namespace EddiEvents
             VARIABLES.Add("faction", "The faction controlling the station at which the commander has docked");
             VARIABLES.Add("factionstate", "The state of the faction controlling the station at which the commander has docked");
             VARIABLES.Add("economy", "The economy of the station at which the commander has docked");
+            VARIABLES.Add("secondeconomy", "The secondary economy of the station at which the commander has docked");
             VARIABLES.Add("government", "The government of the station at which the commander has docked");
             VARIABLES.Add("security", "The security of the station at which the commander has docked");
             VARIABLES.Add("distancefromstar", "The distance of this station from the star (light seconds)");
@@ -37,17 +38,19 @@ namespace EddiEvents
 
         public string state { get; private set; }
 
-        public string model { get; private set; }
+        public string model => stationModel.localizedName;
 
         public string allegiance => Allegiance.localizedName;
 
         public string faction { get; private set; }
 
-        public string factionstate { get; private set; }
+        public string factionstate => factionState.localizedName;
 
-        public string economy { get; private set; }
+        public string economy => economyShares.Count > 0 ? (economyShares[0]?.economy ?? Economy.None).localizedName : Economy.None.localizedName;
 
-        public string government { get; private set; }
+        public string secondeconomy => economyShares.Count > 1 ? (economyShares[1]?.economy ?? Economy.None).localizedName : Economy.None.localizedName;
+
+        public string government => Government.localizedName;
 
         public decimal? distancefromstar { get; private set; }
 
@@ -66,24 +69,26 @@ namespace EddiEvents
 
         // These properties are not intended to be user facing
         public long systemAddress { get; private set; }
-
+        public StationModel stationModel { get; private set; } = StationModel.None;
         public Superpower Allegiance { get; private set; } = Superpower.None;
+        public List<StationService> stationServices { get; private set; } = new List<StationService>();
+        public FactionState factionState { get; private set; } = FactionState.None;
+        public Government Government { get; private set; } = Government.None;
+        public List<EconomyShare> economyShares { get; private set; } = new List<EconomyShare>() { new EconomyShare(Economy.None, 0M), new EconomyShare(Economy.None, 0M) };
 
-        public List<StationService> stationServices { get; private set; }
-
-        public DockedEvent(DateTime timestamp, string system, long systemAddress, long marketId, string station, string state, string model, Superpower allegiance, string faction, FactionState factionstate, Economy economy, Government government, decimal? distancefromstar, List<StationService> stationServices) : base(timestamp, NAME)
+        public DockedEvent(DateTime timestamp, string system, long systemAddress, long marketId, string station, string state, StationModel stationModel, Superpower Allegiance, string faction, FactionState factionState, List<EconomyShare> Economies, Government Government, decimal? distancefromstar, List<StationService> stationServices) : base(timestamp, NAME)
         {
             this.system = system;
             this.systemAddress = systemAddress;
             this.marketId = marketId;
             this.station = station;
             this.state = state;
-            this.model = model;
-            this.Allegiance = (allegiance ?? Superpower.None);
+            this.stationModel = stationModel ?? StationModel.None;
+            this.Allegiance = Allegiance ?? Superpower.None;
             this.faction = faction;
-            this.factionstate = (factionstate ?? FactionState.None).localizedName;
-            this.economy = (economy ?? Economy.None).localizedName;
-            this.government = (government ?? Government.None).localizedName;
+            this.factionState = factionState ?? FactionState.None;
+            this.economyShares = Economies;
+            this.Government = Government ?? Government.None;
             this.distancefromstar = distancefromstar;
             this.stationServices = stationServices;
         }